+++ internal/parsers/tokens.go
@@ -124,7 +124,7 @@ func (p TransferParser) makeFA12Transfers(operation models.Operation, parameters
 	}
 	transfer.From = fromAddr
 	transfer.To = toAddr
-	transfer.Amount = parameters.Get("args.1.args.1.int").Int()
+	transfer.Amount = parameters.Get("args.1.args.1.int").Float()
 	return []*models.Transfer{transfer}, nil
 }
 
@@ -143,7 +143,7 @@ func (p TransferParser) makeFA2Transfers(operation models.Operation, parameters 
 			transfer := models.EmptyTransfer(operation)
 			transfer.From = fromAddr
 			transfer.To = toAddr
-			transfer.Amount = to.Get("args.1.args.1.int").Int()
+			transfer.Amount = to.Get("args.1.args.1.int").Float()
 			transfer.TokenID = to.Get("args.1.args.0.int").Int()
 			transfers = append(transfers, transfer)
 		}
@@ -206,9 +206,9 @@ func (p TransferParser) parseResponse(parser database.BalanceViewParser, operati
 			} else {
 				transfer.From = balance.Address
 			}
-			transfer.Amount = delta
+			transfer.Amount = float64(delta)
 		} else {
-			transfer.Amount = balance.Value
+			transfer.Amount = float64(balance.Value)
 			transfer.To = balance.Address
 		}
 		transfer.TokenID = balance.TokenID
